improve progress bars for uploads

+++ lib/util/progressbar.go
@@ -8,9 +8,9 @@ import (
 )
 
 // Create new progress bar with custom options.
-func NewProgressBar(count int, name string) *mpb.Bar {
+func NewProgressBar(count int64, name string) *mpb.Bar {
 	p := mpb.New()
-	return p.New(int64(count),
+	return p.New(count,
 		mpb.BarStyle().Lbound("[").Filler("=").Tip(">").Padding("-").Rbound("]"),
 		mpb.PrependDecorators(
 			decor.Name(name, decor.WC{W: len(name) + 2, C: decor.DidentRight}),
